+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add `deleteAccount` function.
+- Include js files in dist bundle.
+
 ## 0.1.1-5 (2021-10-08)
 
 - Change `dist` for JS.